NF: Extract helper method from Decks
Only used once
